+++ src/model/qwen3/prefill.rs
@@ -1,0 +1,243 @@
+use anyhow::Result;
+
+use super::weights::{Qwen3Model, TransformerBlock};
+use crate::model::kv_cache::KVCache;
+use crate::ops;
+use crate::tensor::*;
+
+/// Pre-allocated scratch buffers for one prefill forward pass.
+/// Created once per prefill in `process_all_layers_batch`, eliminating
+/// per-layer `cuMemAllocAsync` overhead (~11k calls / 88ms at seq=2048).
+///
+/// Buffer reuse across steps (all kernels serialized on a single stream):
+///   `normed`  reused for `normed2`  (steps 1-4 done before step 8)
+///   `o_buf`   reused for `mlp_out`  (step 7 done before step 12)
+struct PrefillBuffers {
+    /// Output ping-pong: layer writes result here; caller swaps with the incoming hidden.
+    hidden_out: HiddenStates, // hidden_dim × seq_len
+    normed: HiddenStates,      // hidden_dim × seq_len (reused for normed2)
+    q_batch: HiddenStates,     // q_dim × seq_len
+    k_batch: HiddenStates,     // kv_dim × seq_len
+    v_batch: HiddenStates,     // kv_dim × seq_len
+    o_buf: HiddenStates,       // hidden_dim × seq_len (reused for mlp_out)
+    gate_out: HiddenStates,    // inter_dim × seq_len
+    up_out: HiddenStates,      // inter_dim × seq_len
+    act_out: HiddenStates,     // inter_dim × seq_len
+    attn_output: HiddenStates, // q_dim × seq_len
+}
+
+impl PrefillBuffers {
+    fn new(
+        ctx: &DeviceContext,
+        hidden_dim: usize,
+        q_dim: usize,
+        kv_dim: usize,
+        inter_dim: usize,
+        seq_len: usize,
+    ) -> Result<Self> {
+        Ok(Self {
+            hidden_out: HiddenStates::zeros(ctx, hidden_dim, seq_len)?,
+            normed: HiddenStates::zeros(ctx, hidden_dim, seq_len)?,
+            q_batch: HiddenStates::zeros(ctx, q_dim, seq_len)?,
+            k_batch: HiddenStates::zeros(ctx, kv_dim, seq_len)?,
+            v_batch: HiddenStates::zeros(ctx, kv_dim, seq_len)?,
+            o_buf: HiddenStates::zeros(ctx, hidden_dim, seq_len)?,
+            gate_out: HiddenStates::zeros(ctx, inter_dim, seq_len)?,
+            up_out: HiddenStates::zeros(ctx, inter_dim, seq_len)?,
+            act_out: HiddenStates::zeros(ctx, inter_dim, seq_len)?,
+            attn_output: HiddenStates::zeros(ctx, q_dim, seq_len)?,
+        })
+    }
+}
+
+impl Qwen3Model {
+    #[fastrace::trace(name = "get_embeddings_batch")]
+    pub(super) fn get_embeddings_batch(&self, token_ids: &[u32]) -> Result<HiddenStates> {
+        let seq_len = token_ids.len();
+        let hidden_dim = self.config.hidden_size;
+
+        // Copy token IDs to GPU
+        let token_ids_i32: Vec<i32> = token_ids.iter().map(|&x| x as i32).collect();
+        let token_ids_gpu = self
+            .ctx
+            .stream
+            .clone_htod(&token_ids_i32)
+            .map_err(|e| anyhow::anyhow!("H2D copy failed: {}", e))?;
+
+        let mut out = HiddenStates::zeros(&self.ctx, hidden_dim, seq_len)?;
+        ops::embedding_batch(&self.ctx, &self.embed_tokens, &token_ids_gpu, &mut out)?;
+
+        Ok(out)
+    }
+
+    #[fastrace::trace(name = "process_all_layers_batch")]
+    pub(super) fn process_all_layers_batch(
+        &self,
+        mut hidden: HiddenStates,
+        start_pos: usize,
+        kv_cache: &mut KVCache,
+    ) -> Result<HiddenStates> {
+        let seq_len = hidden.seq_len;
+        let num_heads = self.config.num_attention_heads;
+        let num_kv_heads = self.config.num_key_value_heads;
+        let head_dim = self.config.head_dim;
+        let inter_dim = self.config.intermediate_size;
+        let q_dim = num_heads * head_dim;
+        let kv_dim = num_kv_heads * head_dim;
+
+        // Allocate all intermediates once — eliminates ~11k cuMemAllocAsync calls.
+        let mut bufs = PrefillBuffers::new(
+            &self.ctx,
+            self.config.hidden_size,
+            q_dim,
+            kv_dim,
+            inter_dim,
+            seq_len,
+        )?;
+
+        for (layer_idx, layer) in self.layers.iter().enumerate() {
+            self.forward_layer_batch(
+                layer_idx,
+                layer,
+                &mut hidden,
+                start_pos,
+                kv_cache,
+                &mut bufs,
+            )?;
+        }
+
+        // Increment sequence length AFTER all layers processed
+        for _ in 0..seq_len {
+            kv_cache.increment_seq_len();
+        }
+
+        Ok(hidden)
+    }
+
+    pub(super) fn compute_logits_batch(&self, hidden: &HiddenStates) -> Result<DeviceVec> {
+        let last_hidden = ops::extract_vec(&self.ctx, hidden, hidden.seq_len - 1)?;
+        let normed = ops::rms_norm(
+            &self.ctx,
+            &last_hidden,
+            &self.norm,
+            self.config.rms_norm_eps,
+        )?;
+        ops::linear(&self.ctx, &normed, self.output_projection())
+    }
+
+    fn forward_layer_batch(
+        &self,
+        layer_idx: usize,
+        layer: &TransformerBlock,
+        hidden: &mut HiddenStates,
+        start_pos: usize,
+        kv_cache: &mut KVCache,
+        bufs: &mut PrefillBuffers,
+    ) -> Result<()> {
+        let num_heads = self.config.num_attention_heads;
+        let num_kv_heads = self.config.num_key_value_heads;
+        let head_dim = self.config.head_dim;
+
+        kv_cache.init_if_needed(&self.ctx, self.config.head_dim)?;
+
+        // 1. RMSNorm → bufs.normed
+        ops::rms_norm_batch_into(
+            &self.ctx,
+            hidden,
+            &layer.input_layernorm,
+            self.config.rms_norm_eps,
+            &mut bufs.normed,
+        )?;
+
+        // 2. QKV projections → bufs.q_batch, bufs.k_batch, bufs.v_batch
+        ops::gemm_into(
+            &self.ctx,
+            &layer.attention.q_proj,
+            &bufs.normed,
+            &mut bufs.q_batch,
+        )?;
+        ops::gemm_into(
+            &self.ctx,
+            &layer.attention.k_proj,
+            &bufs.normed,
+            &mut bufs.k_batch,
+        )?;
+        ops::gemm_into(
+            &self.ctx,
+            &layer.attention.v_proj,
+            &bufs.normed,
+            &mut bufs.v_batch,
+        )?;
+
+        // 3. FlashAttention-2 (Triton) → bufs.attn_output
+        let (k_cache_layer, v_cache_layer) = kv_cache.get_cache_mut(&self.ctx, layer_idx)?;
+        ops::prefill_attention_batch(
+            &self.ctx,
+            &mut bufs.q_batch,
+            &mut bufs.k_batch,
+            &bufs.v_batch,
+            &layer.attention.q_norm,
+            &layer.attention.k_norm,
+            &self.cos_cache,
+            &self.sin_cache,
+            k_cache_layer,
+            v_cache_layer,
+            &mut bufs.attn_output,
+            num_heads,
+            num_kv_heads,
+            head_dim,
+            start_pos,
+            self.config.rms_norm_eps,
+        )?;
+
+        // 4. O projection → bufs.o_buf (as o_batch)
+        ops::gemm_into(
+            &self.ctx,
+            &layer.attention.o_proj,
+            &bufs.attn_output,
+            &mut bufs.o_buf,
+        )?;
+
+        // 5. Residual add: hidden_in + o_batch → bufs.hidden_out
+        ops::add_batch_into(&self.ctx, hidden, &bufs.o_buf, &mut bufs.hidden_out)?;
+        // Swap: hidden = attn_residual, bufs.hidden_out = old hidden_in (now free)
+        std::mem::swap(hidden, &mut bufs.hidden_out);
+
+        // 6. MLP RMSNorm → bufs.normed (reused for normed2; steps 1-4 are done)
+        ops::rms_norm_batch_into(
+            &self.ctx,
+            hidden,
+            &layer.post_attention_layernorm,
+            self.config.rms_norm_eps,
+            &mut bufs.normed,
+        )?;
+
+        // 7. MLP: gate + up → act → down → bufs.o_buf (reused for mlp_out; step 5 is done)
+        ops::gemm_into(
+            &self.ctx,
+            &layer.mlp.gate_proj,
+            &bufs.normed,
+            &mut bufs.gate_out,
+        )?;
+        ops::gemm_into(
+            &self.ctx,
+            &layer.mlp.up_proj,
+            &bufs.normed,
+            &mut bufs.up_out,
+        )?;
+        ops::silu_mul_batch_into(&self.ctx, &bufs.gate_out, &bufs.up_out, &mut bufs.act_out)?;
+        ops::gemm_into(
+            &self.ctx,
+            &layer.mlp.down_proj,
+            &bufs.act_out,
+            &mut bufs.o_buf,
+        )?;
+
+        // 8. Residual add: attn_residual + mlp_out → bufs.hidden_out (old hidden_in, free to overwrite)
+        ops::add_batch_into(&self.ctx, hidden, &bufs.o_buf, &mut bufs.hidden_out)?;
+        // Swap: hidden = layer output, bufs.hidden_out = attn_residual (free next layer)
+        std::mem::swap(hidden, &mut bufs.hidden_out);
+
+        Ok(())
+    }
+}
